feat: update after webinar add NameParameterRepository logic

+++ src/main/java/ru/yandex/practicum/filmorate/storage/like/JdbcLikeRepository.java
@@ -1,25 +1,32 @@
 package ru.yandex.practicum.filmorate.storage.like;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcOperations;
 import org.springframework.stereotype.Repository;
 
+import java.util.HashMap;
+import java.util.Map;
 
 @Repository
 @RequiredArgsConstructor
 public class JdbcLikeRepository implements LikeRepository {
-    private final JdbcTemplate jdbc;
-
+    private final NamedParameterJdbcOperations jdbc;
 
     @Override
     public void addLike(Long filmId, Long userId) {
-        String sql = "INSERT INTO likes (film_id, user_id) VALUES (?, ?)";
-        jdbc.update(sql, filmId, userId);
+        String sql = "INSERT INTO likes (film_id, user_id) VALUES (:filmId, :userId)";
+        Map<String, Object> params = new HashMap<>();
+        params.put("filmId", filmId);
+        params.put("userId", userId);
+        jdbc.update(sql, params);
     }
 
     @Override
     public void removeLike(Long filmId, Long userId) {
-        String sql = "DELETE FROM likes WHERE film_id = ? AND user_id = ?";
-        jdbc.update(sql, filmId, userId);
+        String sql = "DELETE FROM likes WHERE film_id = :filmId AND user_id = :userId";
+        Map<String, Object> params = new HashMap<>();
+        params.put("filmId", filmId);
+        params.put("userId", userId);
+        jdbc.update(sql, params);
     }
 }
